# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/diagnosa": {
            "get": {
                "description": "Get a list of Diagnosa.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Diagnosa"
                ],
                "summary": "Get all Diagnosa.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Diagnosa"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creating a new Diagnosa.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Diagnosa"
                ],
                "summary": "Create New Diagnosa.",
                "parameters": [
                    {
                        "description": "the body to create a new Diagnosa",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.diagnosaInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Diagnosa"
                        }
                    }
                }
            }
        },
        "/diagnosa/{id}": {
            "get": {
                "description": "Get an Diagnosa by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Diagnosa"
                ],
                "summary": "Get Diagnosa.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Diagnosa id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Diagnosa"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a Diagnosa by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Diagnosa"
                ],
                "summary": "Delete one Diagnosa.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Diagnosa id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Diagnosa by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Diagnosa"
                ],
                "summary": "Update Diagnosa.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Diagnosa id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update age diagnosa category",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.diagnosaInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Diagnosa"
                        }
                    }
                }
            }
        },
        "/dokter": {
            "get": {
                "description": "Get a list of Dokter.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dokter"
                ],
                "summary": "Get all Dokter.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Dokter"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creating a new Dokter.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dokter"
                ],
                "summary": "Create New Dokter.",
                "parameters": [
                    {
                        "description": "the body to create a new Dokter",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.dokterInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Dokter"
                        }
                    }
                }
            }
        },
        "/dokter/{id}": {
            "get": {
                "description": "Get an Dokter by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dokter"
                ],
                "summary": "Get Dokter.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dokter id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Dokter"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a Dokter by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dokter"
                ],
                "summary": "Delete one Dokter.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dokter id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Dokter by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dokter"
                ],
                "summary": "Update Dokter.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dokter id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update age dokter category",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.dokterInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Dokter"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.diagnosaInput": {
            "type": "object",
            "properties": {
                "berat_badan": {
                    "type": "number"
                },
                "dokterID": {
                    "type": "integer"
                },
                "hasil_diagnosa": {
                    "type": "string"
                },
                "keluhan": {
                    "type": "string"
                },
                "pasienID": {
                    "type": "integer"
                },
                "tanggal": {
                    "type": "string"
                },
                "tekanan_darah": {
                    "type": "string"
                },
                "tindakan": {
                    "type": "string"
                }
            }
        },
        "controller.dokterInput": {
            "type": "object",
            "properties": {
                "jadwal": {
                    "type": "string"
                },
                "jenis_kelamin": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "no_hp": {
                    "type": "string"
                },
                "spesialis": {
                    "type": "string"
                },
                "tanggal_lahir": {
                    "type": "string"
                }
            }
        },
...
        "models.Diagnosa": {
            "type": "object",
            "properties": {
                "berat_badan": {
                    "type": "number"
                },
                "dokterID": {
                    "type": "integer"
                },
                "hasil_diagnosa": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "keluhan": {
                    "type": "string"
                },
                "pasienID": {
                    "type": "integer"
                },
                "tanggal": {
                    "type": "string"
                },
                "tekanan_darah": {
                    "type": "string"
                },
                "tindakan": {
                    "type": "string"
                }
            }
        },
        "models.Dokter": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "jadwal": {
                    "type": "string"
                },
                "jenis_kelamin": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "no_hp": {
                    "type": "string"
                },
                "spesialis": {
                    "type": "string"
                },
                "tanggal_lahir": {
                    "type": "string"
                }
            }
        },
...
    }
}